tests: fix segfault on sample_regs_intr
it was possibl that when we called PERF_IOC_DISABLE that a signal
would be generated after it started disabling but before the
signal handler was disabled, and the signal handler would refresh
and re-enable the event, and then we'd munmap and then segfault.

So explicitly disable the signal handler before disabling.

We only saw this on AMD fam15h machines for some reason.