fuzzer: Supply mode argument to open() when required to
When FORTIFY_SOURCE is enabled, libc will terminate the process on
a call to open() with flags that require a mode argument but it's
not provided. Here is the relevant part of the glibc docs:

  The mode argument must be supplied if O_CREAT or O_TMPFILE is
  specified in flags; if it is not supplied, some arbitrary
  bytes from the stack will be applied as the file mode.

And this is what it looks like when running the fuzzer:

  ==================================================
  Starting fuzzing at 2022-03-17 13:15:53
  ==================================================
  *** invalid open call: O_CREAT or O_TMPFILE without mode ***: ./perf_fuzzer terminated
  Aborted

Fix this by supplying a mode argument when the random flags indicate
it's required. The actual mode is not important because files can't
be created in sysfs.

Signed-off-by: James Clark <james.clark@arm.com>